make fingerprint script nice
